Teasing out cmr pieces from granule.js
